KYRA: Limit number of MIDI msgs during fade-out
When a fade-out is performed, a lot of MIDI messages were generated in
a short space of time (over 1500 in a second). Older MIDI modules have
problems dealing with so many messages. My SC-55mkII slows the music
down during the fade-out, and then speeds up the beginning of the next
track as it catches up.

I've fixed this by limiting the fade-out to 25 volume steps. This does
not result in an audible difference in the fade-out, but it does reduce
the number of messages and fixes the timing problems on the SC-55.