KYRA/MIDI: Fix MT-32 overflow on game launch
Several KYRA engine games (Kyrandia 1 & 2, Lands of Lore) would cause
a buffer overflow on the MT-32 when launching the game. This was
caused by the game sending a reset immediately followed by several
more sysexes. The reset did not have enough time to execute, causing
the other sysexes to overflow the buffer.

I've fixed this by using MidiDriver::sendMT32Reset instead of
sendSysEx, which already uses a longer delay, and then increasing the
delay to 130 ms (120 seems to be enough, but I gave it a little extra
to be on the safe side).