KYRA: Initialize default pitch bend and program
During initialization the default values for the pitch bend and
program were sent out to the MIDI device, but were not set in the
internal state of the program. This would prevent the pitch bend and
program values from being reset when unlocking a channel in some
cases, specifically in Legend of Kyrandia when playing a sound effect
on channels 8 or 9. I've fixed this.