KYRA: Apply MIDI control changes to channel state
When a control change event was sent to the MIDI output, it would
store this in the source state and send it out to the MIDI device,
but it would not update the channel state. This would cause wrong
controller values to be used sometimes.

I.e. in Legend of Kyrandia, when Brandon talks to the tree in the
beginning of the game, several music channels are used to play the
tree creaking sound effect. When the channels were unlocked, the
volume was reset to an incorrect value, and the music that resumed
playing on these channels was too loud.

I've added an update of the channel state when receiving a CC.