LURE: Implement TrashReverb
The original interpreter sends out a SysEx to reduce the MT-32 reverb at
several events, including opening a door while outside. I've added sending of
this SysEx to ScummVM.

There are still some things unclear about this function, such as the other
events which trigger calling this function, and what it is supposed to
accomplish.